Make ScalingScrollView generic so it can be used for meets and home page (and other collections)
